src/requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile src/requirements/requirements.in
#
coverage-badge==1.0.1
dj-database-url==0.5.0
django==2.2.4
djangorestframework==3.10.2
mysqlclient==1.4.4
python-decouple==3.1
pytz==2019.2              # via django
sqlparse==0.3.0           # via django
